# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for management students database",
        "title": "Zerg Team Student Information Service",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/homeworks": {
            "get": {
                "description": "Get list of all homeworks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homeworks"
                ],
                "summary": "Get homeworks list",
                "operationId": "homeworks-list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Homework"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    },
                    "default": {
                        "description": "",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create new homework in database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homeworks"
                ],
                "summary": "Create homework",
                "operationId": "create-homework",
                "parameters": [
                    {
                        "description": "homework data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Homework"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    },
                    "default": {
                        "description": "",
                        "schema": {}
                    }
                }
            }
        },
        "/homeworks/:id": {
            "get": {
                "description": "Get homework from database by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homeworks"
                ],
                "summary": "Get homework by id",
                "operationId": "get-homework-by-id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Homework"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    },
                    "default": {
                        "description": "",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Remove homework from database by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homeworks"
                ],
                "summary": "Remove homework",
                "operationId": "remove-homework-by-id",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    },
                    "default": {
                        "description": "",
                        "schema": {}
                    }
                }
            },
            "patch": {
                "description": "Update homework data in database by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "homeworks"
                ],
                "summary": "Update homework",
                "operationId": "update-homework-by-id",
                "parameters": [
                    {
                        "description": "homework data to update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Homework"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    },
                    "default": {
                        "description": "",
                        "schema": {}
                    }
                }
            }
        },
        "/interviews": {
            "get": {
                "description": "Get list of all interviews",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interviews"
                ],
                "summary": "Get interviews list",
                "operationId": "interviews-list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Interview"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    },
                    "default": {
                        "description": "",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create new interview in database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interviews"
                ],
                "summary": "Create interview",
                "operationId": "create-interview",
                "parameters": [
                    {
                        "description": "interview data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Interview"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    },
                    "default": {
                        "description": "",
                        "schema": {}
                    }
                }
            }
        },
        "/interviews/:id": {
            "get": {
                "description": "Get interview from database by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interviews"
                ],
                "summary": "Get interview by id",
                "operationId": "get-interview-by-id",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    },
                    "default": {
                        "description": "",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "description": "Remove interview from database by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interviews"
                ],
                "summary": "Remove interview",
                "operationId": "remove-interview-by-id",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    },
                    "default": {
                        "description": "",
                        "schema": {}
                    }
                }
            },
            "patch": {
                "description": "Update interview data in database by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "interviews"
                ],
                "summary": "Update interview",
                "operationId": "update-interview-by-id",
                "parameters": [
                    {
                        "description": "interview data to update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Interview"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    },
                    "default": {
                        "description": "",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Homework": {
            "type": "object",
            "properties": {
                "deadline": {
                    "type": "string"
                },
                "homeworkID": {
                    "type": "integer"
                },
                "text": {
                    "type": "string"
                }
            }
        },
...
        "models.Interview": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "mark": {
                    "type": "integer"
                },
                "mentorID": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "studentID": {
                    "type": "integer"
                }
            }
        },
...
    }
}